Code refactoring to match Arik's new controllers/handlers organization scheme
