+++ redash/devspark/custom_tasks/__init__.py
@@ -1,0 +1,1 @@
+from redash.devspark.custom_tasks import report_tasks
